cleaned up naming and file names a bit too
